[허지웅] sprint3

+++ src/main/java/com/sprint/mission/discodeit/repository/file/FileChannelRepository.java
@@ -1,9 +1,13 @@
 package com.sprint.mission.discodeit.repository.file;
 
-import com.sprint.mission.discodeit.entity.BaseChannel;
+import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.exception.ChannelNotFoundException;
+import com.sprint.mission.discodeit.repository.AbstractFileRepository;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.util.FileUtil;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.stereotype.Repository;
 
 import java.io.File;
 import java.util.List;
@@ -11,72 +15,60 @@ import java.util.Optional;
 
 import static com.sprint.mission.discodeit.constant.FileConstant.CHANNEL_FILE;
 
-public class FileChannelRepository implements ChannelRepository {
+@Repository
+@ConditionalOnProperty(name = "app.repository.type", havingValue = "file")
+public class FileChannelRepository extends AbstractFileRepository<Channel> implements ChannelRepository{
 
-  private static FileChannelRepository instance;
-
-  private FileChannelRepository() {
-  }
-
-  public static FileChannelRepository getInstance() {
-    if (instance == null) {
-      synchronized (FileChannelRepository.class) {
-        if (instance == null) {
-          instance = new FileChannelRepository();
-        }
-      }
-    }
-    return instance;
+  public FileChannelRepository(@Value("${app.file.channel-file}") String filePath) {
+    super(filePath);
   }
 
   @Override
-  public BaseChannel save(BaseChannel baseChannel) {
-    List<BaseChannel> channels = FileUtil.loadAllFromFile(CHANNEL_FILE, BaseChannel.class);
-    channels.add(baseChannel);
-    FileUtil.saveAllToFile(CHANNEL_FILE, channels);
-    return baseChannel;
+  public Channel save(Channel channel) {
+    List<Channel> channels = loadAll(Channel.class);
+    channels.add(channel);
+    saveAll(channels);
+    return channel;
   }
 
   @Override
-  public Optional<BaseChannel> findById(String id) {
-    List<BaseChannel> channels = FileUtil.loadAllFromFile(CHANNEL_FILE, BaseChannel.class);
+  public Optional<Channel> findById(String id) {
+    List<Channel> channels = loadAll(Channel.class);
     return channels.stream().filter(c -> c.getUUID().equals(id)).findAny();
   }
 
   @Override
-  public List<BaseChannel> findAll() {
-    return FileUtil.loadAllFromFile(CHANNEL_FILE, BaseChannel.class);
+  public List<Channel> findAll() {
+    return loadAll(Channel.class);
   }
 
   @Override
-  public BaseChannel update(BaseChannel channel) {
-    List<BaseChannel> channels = FileUtil.loadAllFromFile(CHANNEL_FILE, BaseChannel.class);
-    BaseChannel targetChannel = channels.stream()
+  public Channel update(Channel channel) {
+    List<Channel> channels = loadAll(Channel.class);
+    Channel targetChannel = channels.stream()
         .filter(c -> c.getUUID().equals(channel.getUUID()))
         .findAny()
         .orElseThrow(() -> new ChannelNotFoundException());
     channels.remove(targetChannel);
     channels.add(channel);
-    FileUtil.saveAllToFile(CHANNEL_FILE, channels);
+    saveAll(channels);
     return channel;
   }
 
   @Override
   public void delete(String id) {
-    List<BaseChannel> channels = FileUtil.loadAllFromFile(CHANNEL_FILE, BaseChannel.class);
-    BaseChannel targetChannel = channels.stream()
+    List<Channel> channels = loadAll(Channel.class);
+    Channel targetChannel = channels.stream()
         .filter(c -> c.getUUID().equals(id))
         .findAny()
         .orElseThrow(() -> new ChannelNotFoundException());
     channels.remove(targetChannel);
-    FileUtil.saveAllToFile(CHANNEL_FILE, channels);
+    saveAll(channels);
   }
 
   @Override
   public void clear() {
-    File file = new File(CHANNEL_FILE);
-    if (file.exists()) {
-      file.delete();
-    }
+    File file = new File(getFilePath());
+    if(file.exists()) file.delete();
   }
 }
