+++ src/main/java/com/sprint/mission/discodeit/repository/file/FileUserRepository.java
@@ -1,84 +1,86 @@
 package com.sprint.mission.discodeit.repository.file;
 
+import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.exception.UserNotFoundException;
+import com.sprint.mission.discodeit.repository.AbstractFileRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.util.FileUtil;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.stereotype.Repository;
 
 import java.io.File;
 import java.util.List;
 import java.util.Optional;
 
-import static com.sprint.mission.discodeit.constant.FileConstant.CHANNEL_FILE;
 import static com.sprint.mission.discodeit.constant.FileConstant.USER_FILE;
 
-public class FileUserRepository implements UserRepository {
+@Repository
+@ConditionalOnProperty(name = "app.repository.type", havingValue = "file")
+public class FileUserRepository extends AbstractFileRepository<User> implements UserRepository{
 
-  private static FileUserRepository instance;
-
-  private FileUserRepository() {
-  }
-
-  public static FileUserRepository getInstance() {
-    if (instance == null) {
-      synchronized (FileUserRepository.class) {
-        if (instance == null) {
-          instance = new FileUserRepository();
-        }
-      }
-    }
-    return instance;
+  public FileUserRepository(@Value("${app.file.user-file}") String filePath) {
+    super(filePath);
   }
 
   @Override
   public User create(User user) {
-    List<User> users = FileUtil.loadAllFromFile(USER_FILE, User.class);
+    List<User> users = loadAll(User.class);
+
     users.add(user);
-    FileUtil.saveAllToFile(USER_FILE, users);
+    saveAll(users);
+
     return user;
   }
 
   @Override
   public Optional<User> findById(String id) {
-    List<User> users = FileUtil.loadAllFromFile(USER_FILE, User.class);
+    List<User> users = loadAll(User.class);
     return users.stream().filter(u -> u.getUUID().equals(id)).findAny();
   }
 
   @Override
+  public Optional<User> findByUsername(String username) {
+    List<User> users = loadAll(User.class);
+    return users.stream().filter(u -> u.getUsername().equals(username)).findAny();
+  }
+
+  @Override
   public List<User> findAll() {
-    return FileUtil.loadAllFromFile(USER_FILE, User.class);
+    return loadAll(User.class);
   }
 
   @Override
   public User update(User user) {
-    List<User> users = FileUtil.loadAllFromFile(USER_FILE, User.class);
+    List<User> users = loadAll(User.class);
+
     User targetUser = users.stream()
         .filter(u -> u.getUUID().equals(user.getUUID()))
         .findAny()
         .orElseThrow(() -> new UserNotFoundException());
+
     users.remove(targetUser);
     users.add(user);
-    FileUtil.saveAllToFile(USER_FILE, users);
+    saveAll(users);
 
     return user;
   }
 
   @Override
   public void delete(String userId) {
-    List<User> users = FileUtil.loadAllFromFile(USER_FILE, User.class);
+    List<User> users = loadAll(User.class);
     User targetUser = users.stream()
         .filter(u -> u.getUUID().equals(userId))
         .findAny()
         .orElseThrow(() -> new UserNotFoundException());
     users.remove(targetUser);
-    FileUtil.saveAllToFile(USER_FILE, users);
+    saveAll(users);
   }
 
   @Override
   public void clear() {
-    File file = new File(USER_FILE);
-    if (file.exists()) {
-      file.delete();
-    }
+    File file = new File(getFilePath());
+    if(file.exists()) file.delete();
   }
 }
